+++ indexer/src/cache/mod.rs
@@ -2,8 +2,7 @@ pub mod redis;
 use anyhow::Error;
 use async_trait::async_trait;
 
-use crate::consumer::kakfa_message::UserOpPolicyData;
-
+use crate::model::user_op_policy::UserOpPolicyData;
 #[async_trait]
 pub trait Cache {
     async fn update_userop_policy(&self, user_op_hash: &str, partial: UserOpPolicyData) -> Result<(), Error>;
