+++ indexer/src/consumer/kakfa_message.rs
@@ -21,17 +21,6 @@ pub struct UserOpMessage {
     pub enabled_limits: Option<Vec<String>>,
 }
 
-#[derive(Debug, Serialize, Deserialize, Default)]
-pub struct UserOpPolicyData {
-    pub policy_id: Option<String>,
-    pub native_usd_price: Option<String>,
-    pub actual_gas_cost: Option<String>,
-    pub actual_gas_used: Option<String>,
-    pub sender: Option<String>,
-    pub enabled_limits: Option<Vec<String>>,
-}
-
-
 #[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize, Display)]
 pub enum Status {
     Failed,
