+++ indexer/src/lib.rs
@@ -3,3 +3,4 @@ pub mod consumer;
 pub mod storage;
 pub mod app;
 pub mod cache;
+pub mod model;
